Update dice test package name and format classes
