+++ src/main/java/edu/ntnu/idi/idatt/boardgame/App.java
@@ -11,7 +11,11 @@ public class App extends Application {
   public void start(Stage stage) {
     GameController gameController = new GameController();
 
-    VBox root = new VBox(gameController.getGameBoard(), gameController.getRollDiceButton(), gameController.getLogLabel());
+    VBox root =
+        new VBox(
+            gameController.getGameBoard(),
+            gameController.getRollDiceButton(),
+            gameController.getLogLabel());
     Scene scene = new Scene(root, 700, 800);
     stage.setScene(scene);
     stage.setTitle("JavaFX Maven App");
